refactor(api): migrate from TypeScript to JavaScript with JSDoc types
Convert the `api/` package sources from TypeScript to JavaScript, keeping
`tsc` for strict type checking via JSDoc annotations (`@import` for
external types, `@typedef` for derived types).

- `tsconfig.json` uses `allowJs`/`checkJs`/`emitDeclarationOnly` to
  type-check the `.js` sources and emit `dist/index.d.ts`.
- Run the generator with `node cli.js` instead of `ts-node`.
- Add a `files` allowlist so the published tarball still includes the
  generated `dist/index.d.ts`, which is no longer force-included once
  `main` points outside `dist/`.
- Keep `typescript-eslint` for linting and add an `env` so `no-undef`
  resolves globals like `URL` in the `.js` sources.

The package now ships ESM (`"type": "module"`) instead of CommonJS.
