+++ package.json
@@ -25,6 +25,10 @@
     ]
   },
   "eslintConfig": {
+    "env": {
+      "es2024": true,
+      "node": true
+    },
     "parser": "@typescript-eslint/parser",
     "plugins": [
       "@typescript-eslint"
